* rewrote code
