+++ homeworks/andrew-kondratiyk/20-react-api/src/components/common/edit-todo-card/EditTodoCard.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import {TodoContext} from '@/context/TodoContext';
-import {TodoResponse} from '@/types/todos';
+import {TodoContext, TodoResponse} from '@/context/TodoContext';
 import {useContext, useState} from 'react';
 
 import styles from './EditTodoCard.module.css';
