+++ homeworks/andrew-kondratiyk/20-react-api/src/components/common/todo-list/TodoList.tsx
@@ -1,16 +1,19 @@
 import TodoCard from '@/components/common/todo-card/TodoCard';
-import {TodoContext} from '@/context/TodoContext';
-import {TodoResponse} from '@/types/todos';
-import {useContext} from 'react';
+import {TodoContext, TodoResponse} from '@/context/TodoContext';
+import {useContext, useMemo} from 'react';
 
 const TodoList = () => {
   const {todos} = useContext(TodoContext);
 
   return (
     <>
-      {todos?.map((todo: TodoResponse) => (
-        <TodoCard todo={todo} key={todo.id} />
-      ))}
+      {useMemo(
+        () =>
+          todos?.map((todo: TodoResponse) => (
+            <TodoCard todo={todo} key={todo.id} />
+          )),
+        [todos],
+      )}
     </>
   );
 };
