+++ homeworks/andrew-kondratiyk/20-react-api/src/components/pages/main-page/MainPage.tsx
@@ -1,27 +1,30 @@
 'use client';
 
-import EditTodoCard from '@/components/common/edit-todo-card/EditTodoCard';
+import AddTodoContainer from '@/components/common/add-todo-container/AddTodoContainer';
 import Header from '@/components/common/header/Header';
 import TodoList from '@/components/common/todo-list/TodoList';
 import {TodoContext, todoReducer} from '@/context/TodoContext';
-import {useEffect, useReducer, useState} from 'react';
+import {useEffect, useReducer} from 'react';
 import styles from './MainPage.module.css';
 
 const MainPage = () => {
-  const [isAddTodo, setIsAddTodo] = useState(false);
-  const [todos, dispatch] = useReducer(todoReducer, []);
-
-  useEffect(() => {
+  const [todos, dispatchTodos] = useReducer(todoReducer, [], () => {
+    if (!window) return [];
     const todosLS = window.localStorage.getItem('todos');
-    dispatch({type: 'init', todos: todosLS ? JSON.parse(todosLS) : []});
-  }, []);
+    return todosLS ? JSON.parse(todosLS) : [];
+  });
 
-  useEffect(() => {
-    localStorage.setItem('todos', JSON.stringify(todos));
-  }, [todos]);
+  // useEffect(() => {
+  //   const todosLS = window.localStorage.getItem('todos');
+  //   dispatchTodos({type: 'init', todos: todosLS ? JSON.parse(todosLS) : []});
+  // }, []);
+  //
+  // useEffect(() => {
+  //   localStorage.setItem('todos', JSON.stringify(todos));
+  // }, [todos]);
 
   return (
-    <TodoContext.Provider value={{todos, dispatchTodos: dispatch}}>
+    <TodoContext.Provider value={{todos, dispatchTodos}}>
       <main className={styles.container}>
         <Header />
         <div className={styles.divider}></div>
@@ -29,16 +32,7 @@ const MainPage = () => {
           <TodoList />
         </div>
         <div className={styles.addTaskContainer}>
-          {isAddTodo ? (
-            <EditTodoCard onCancel={() => setIsAddTodo(false)} />
-          ) : (
-            <div
-              onClick={() => setIsAddTodo(true)}
-              className={styles.addTodoBtn}
-            >
-              Add Task
-            </div>
-          )}
+          <AddTodoContainer />
         </div>
       </main>
     </TodoContext.Provider>
