--- homeworks/andrew-kondratiyk/20-react-api/src/types/todos.ts
@@ -1,5 +1,0 @@
-export type TodoResponse = {
-  id: string;
-  title: string;
-  completed: boolean;
-};
